refactor: extract helper to read input secrets
